// frontend/types/contentTypes.d.ts
export interface ApiClassClass extends Struct.CollectionTypeSchema {
  collectionName: 'classes';
  info: {
    singularName: 'class';
    pluralName: 'classes';
    displayName: 'Class';
    description: '';
  };
  options: {
    draftAndPublish: false;
  };
  attributes: {
    name: Schema.Attribute.String & Schema.Attribute.Required;
    description: Schema.Attribute.Text;
    logo: Schema.Attribute.Media<'images' | 'files' | 'videos' | 'audios'>;
    skills: Schema.Attribute.Relation<'oneToMany', 'api::skill.skill'>;
    slug: Schema.Attribute.UID<'name'> & Schema.Attribute.Required;
    strength: Schema.Attribute.Integer;
    agility: Schema.Attribute.Integer;
    stamina: Schema.Attribute.Integer;
    intelligence: Schema.Attribute.Integer;
    volition: Schema.Attribute.Integer;
    faith: Schema.Attribute.Integer;
    createdAt: Schema.Attribute.DateTime;
    updatedAt: Schema.Attribute.DateTime;
    publishedAt: Schema.Attribute.DateTime;
    createdBy: Schema.Attribute.Relation<'oneToOne', 'admin::user'> &
      Schema.Attribute.Private;
    updatedBy: Schema.Attribute.Relation<'oneToOne', 'admin::user'> &
      Schema.Attribute.Private;
    locale: Schema.Attribute.String;
  };
}